requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-ipware==6.0.4
frozenlist==1.4.1
geoip2==4.8.0
idna==3.6
maxminddb==2.6.0
multidict==6.0.5
pillow==10.2.0
python-ipware==2.0.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
